backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
joblib
pandas
google-genai
